move mtgoparser integration tests
